A bit of consolidation and more fiddling with retries.
