+++ Solutions/Marain.Workflow.Api.Specs/Bindings/WorkflowStoreCleanupBindings.cs
@@ -15,6 +15,7 @@ namespace Marain.Workflows.Api.Specs.Bindings
     using Marain.TenantManagement.Testing;
     using Marain.Workflows;
     using Marain.Workflows.Api.Specs.Steps;
+    using Marain.Workflows.Internal;
     using Marain.Workflows.Storage;
     using Microsoft.Extensions.DependencyInjection;
     using TechTalk.SpecFlow;
@@ -49,21 +50,21 @@ namespace Marain.Workflows.Api.Specs.Bindings
                     {
                         ITenantedWorkflowStoreFactory workflowStoreFactory = serviceProvider.GetRequiredService<ITenantedWorkflowStoreFactory>();
                         var workflowStore = (CosmosWorkflowStore)await workflowStoreFactory.GetWorkflowStoreForTenantAsync(transientTenant).ConfigureAwait(false);
-                        await WorkflowRetryHelper.ExecuteWithStandardTestRetryRulesAsync(() => workflowStore.Container.DeleteContainerAsync()).ConfigureAwait(false);
+                        await WorkflowRetryHelper.ExecuteWithRetryRulesAsync(() => workflowStore.Container.DeleteContainerAsync()).ConfigureAwait(false);
                     }).ConfigureAwait(false);
 
                 await context.RunAndStoreExceptionsAsync(async () =>
                 {
                     ITenantedWorkflowInstanceStoreFactory workflowInstanceStoreFactory = serviceProvider.GetRequiredService<ITenantedWorkflowInstanceStoreFactory>();
                     var workflowInstanceStore = (CosmosWorkflowInstanceStore)await workflowInstanceStoreFactory.GetWorkflowInstanceStoreForTenantAsync(transientTenant).ConfigureAwait(false);
-                    await WorkflowRetryHelper.ExecuteWithStandardTestRetryRulesAsync(() => workflowInstanceStore.Container.DeleteContainerAsync()).ConfigureAwait(false);
+                    await WorkflowRetryHelper.ExecuteWithRetryRulesAsync(() => workflowInstanceStore.Container.DeleteContainerAsync()).ConfigureAwait(false);
                 }).ConfigureAwait(false);
 
                 await context.RunAndStoreExceptionsAsync(async () =>
                 {
                     ITenantedWorkflowInstanceChangeLogFactory workflowInstanceChangeLogFactory = serviceProvider.GetRequiredService<ITenantedWorkflowInstanceChangeLogFactory>();
                     var workflowInstanceChangeLog = (CosmosWorkflowInstanceChangeLog)await workflowInstanceChangeLogFactory.GetWorkflowInstanceChangeLogWriterForTenantAsync(transientTenant).ConfigureAwait(false);
-                    await WorkflowRetryHelper.ExecuteWithStandardTestRetryRulesAsync(() => workflowInstanceChangeLog.Container.DeleteContainerAsync()).ConfigureAwait(false);
+                    await WorkflowRetryHelper.ExecuteWithRetryRulesAsync(() => workflowInstanceChangeLog.Container.DeleteContainerAsync()).ConfigureAwait(false);
                 }).ConfigureAwait(false);
             }
         }
