+++ Solutions/Marain.Workflow.Specs/Bindings/WorkflowCosmosDbBindings.cs
@@ -13,6 +13,7 @@ namespace Marain.Workflows.Specs.Bindings
     using Corvus.Retry.Strategies;
     using Corvus.Tenancy;
     using Corvus.Testing.SpecFlow;
+    using Marain.Workflows.Internal;
     using Marain.Workflows.Specs.Steps;
     using Marain.Workflows.Storage;
     using Microsoft.Azure.Cosmos;
@@ -74,7 +75,7 @@ namespace Marain.Workflows.Specs.Bindings
                 testDocumentRepositoryContainerDefinition,
                 cosmosConfig)).ConfigureAwait(false);
 
-            Container testDocumentsRepository = await WorkflowRetryHelper.ExecuteWithStandardTestRetryRulesAsync(
+            Container testDocumentsRepository = await WorkflowRetryHelper.ExecuteWithRetryRulesAsync(
                 () => factory.GetContainerForTenantAsync(
                     testTenant,
                     testDocumentRepositoryContainerDefinition)).ConfigureAwait(false);
@@ -100,22 +101,22 @@ namespace Marain.Workflows.Specs.Bindings
             var workflowStore = (CosmosWorkflowStore)await workflowStoreFactory.GetWorkflowStoreForTenantAsync(tenant).ConfigureAwait(false);
 
             await featureContext.RunAndStoreExceptionsAsync(
-                () => WorkflowRetryHelper.ExecuteWithStandardTestRetryRulesAsync(() => workflowStore.Container.DeleteContainerAsync())).ConfigureAwait(false);
+                () => WorkflowRetryHelper.ExecuteWithRetryRulesAsync(() => workflowStore.Container.DeleteContainerAsync())).ConfigureAwait(false);
 
             ITenantedWorkflowInstanceStoreFactory workflowInstanceStoreFactory = serviceProvider.GetRequiredService<ITenantedWorkflowInstanceStoreFactory>();
             var workflowInstanceStore = (CosmosWorkflowInstanceStore)await workflowInstanceStoreFactory.GetWorkflowInstanceStoreForTenantAsync(tenant).ConfigureAwait(false);
 
             await featureContext.RunAndStoreExceptionsAsync(
-                () => WorkflowRetryHelper.ExecuteWithStandardTestRetryRulesAsync(() => workflowInstanceStore.Container.DeleteContainerAsync())).ConfigureAwait(false);
+                () => WorkflowRetryHelper.ExecuteWithRetryRulesAsync(() => workflowInstanceStore.Container.DeleteContainerAsync())).ConfigureAwait(false);
 
             ITenantedWorkflowInstanceChangeLogFactory workflowInstanceChangeLogFactory = serviceProvider.GetRequiredService<ITenantedWorkflowInstanceChangeLogFactory>();
             var workflowInstanceChangeLog = (CosmosWorkflowInstanceChangeLog)await workflowInstanceChangeLogFactory.GetWorkflowInstanceChangeLogWriterForTenantAsync(tenant).ConfigureAwait(false);
 
             await featureContext.RunAndStoreExceptionsAsync(
-                () => WorkflowRetryHelper.ExecuteWithStandardTestRetryRulesAsync(() => workflowInstanceChangeLog.Container.DeleteContainerAsync())).ConfigureAwait(false);
+                () => WorkflowRetryHelper.ExecuteWithRetryRulesAsync(() => workflowInstanceChangeLog.Container.DeleteContainerAsync())).ConfigureAwait(false);
 
             await featureContext.RunAndStoreExceptionsAsync(
-                () => WorkflowRetryHelper.ExecuteWithStandardTestRetryRulesAsync(() => featureContext.Get<Container>(TestDocumentsRepository).DeleteContainerAsync())).ConfigureAwait(false);
+                () => WorkflowRetryHelper.ExecuteWithRetryRulesAsync(() => featureContext.Get<Container>(TestDocumentsRepository).DeleteContainerAsync())).ConfigureAwait(false);
         }
     }
 }
