+++ Solutions/Marain.Workflow.Specs/Steps/InvokeExternalServiceActionSteps.cs
@@ -11,6 +11,7 @@ namespace Marain.Workflows.Specs.Steps
     using Corvus.Identity.ManagedServiceIdentity.ClientAuthentication;
     using Corvus.Tenancy;
     using Corvus.Testing.SpecFlow;
+    using Marain.Workflows.Internal;
     using Marain.Workflows.Specs.Bindings;
     using Marain.Workflows.Specs.TestObjects;
     using Marain.Workflows.Specs.TestObjects.Triggers;
@@ -62,7 +63,7 @@ namespace Marain.Workflows.Specs.Steps
 
             IWorkflowStore store = await storeFactory.GetWorkflowStoreForTenantAsync(tenant).ConfigureAwait(false);
 
-            await WorkflowRetryHelper.ExecuteWithStandardTestRetryRulesAsync(async () =>
+            await WorkflowRetryHelper.ExecuteWithRetryRulesAsync(async () =>
             {
                 try
                 {
