+++ Solutions/Marain.Workflow.Tenancy.AzureCosmos/Marain/Workflows/TenantedCosmosWorkflowInstanceChangeLogFactory.cs
@@ -42,7 +42,7 @@ namespace Marain.Workflows
         /// <inheritdoc/>
         public async Task<IWorkflowInstanceChangeLogReader> GetWorkflowInstanceChangeLogReaderForTenantAsync(ITenant tenant)
         {
-            Container container = await Retriable.RetryAsync(() => this.containerFactory.GetContainerForTenantAsync(tenant, this.containerDefinition), CancellationToken.None, new Linear(TimeSpan.FromSeconds(20), 5), RetryOnCosmosRequestRateExceededPolicy.Instance).ConfigureAwait(false);
+            Container container = await WorkflowRetryHelper.ExecuteWithRetryRulesAsync(() => this.containerFactory.GetContainerForTenantAsync(tenant, this.containerDefinition)).ConfigureAwait(false);
 
             // No need to cache these instances as they are lightweight wrappers around the container.
             return new CosmosWorkflowInstanceChangeLog(container);
@@ -51,7 +51,7 @@ namespace Marain.Workflows
         /// <inheritdoc/>
         public async Task<IWorkflowInstanceChangeLogWriter> GetWorkflowInstanceChangeLogWriterForTenantAsync(ITenant tenant)
         {
-            Container container = await Retriable.RetryAsync(() => this.containerFactory.GetContainerForTenantAsync(tenant, this.containerDefinition), CancellationToken.None, new Linear(TimeSpan.FromSeconds(20), 5), RetryOnCosmosRequestRateExceededPolicy.Instance).ConfigureAwait(false);
+            Container container = await WorkflowRetryHelper.ExecuteWithRetryRulesAsync(() => this.containerFactory.GetContainerForTenantAsync(tenant, this.containerDefinition)).ConfigureAwait(false);
 
             // No need to cache these instances as they are lightweight wrappers around the container.
             return new CosmosWorkflowInstanceChangeLog(container);
